Iterate over test reader/writer hashes for better fails
Rather than testing the equivilance of the reader and writer constants
with test hashes, we iterate over the key-value pairs to test the
equivilance of the value when the key is called.

This approach has the advantage of drawing attention to specific
discrepencies between the test objects and the reader/writer
constants.